Added layout.html, static, and django-bootstrap
Added layout.html, which can be used as a basic layout for html pages and includes a universal navigation bar. Added static, which allows for the use of images, css, and javascript for html pages. Also added django-bootstrap to requirements.txt and adjusted backports.zoneinfo (its only needed for older versions of python apparently)